requirements.txt
fastapi>=0.122.0
uvicorn>=0.38.0
redis>=7.1.0
pydantic>=2.12.5
python-dotenv==1.2.1
